messy-migration/requirements.txt
Flask==2.3.2
Werkzeug==2.3.6
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.2
python-dotenv==1.0.0
pytest==7.4.0